Moved Column, ForeignKey, Constraint, and Index to database module and removed duplicates
