+++ log4rs.yml
@@ -1,33 +1,11 @@
 # Scan this file for changes every 30 seconds
 refresh_rate: 30 seconds
-
 appenders:
-  # An appender named "stdout" that writes to stdout
-#  stdout:
-#    kind: console
-
   # An appender named "requests" that writes to a file with a custom pattern encoder
-  requests:
+  stdout:
     kind: file
-    path: "log/requests.log"
-#    encoder:
-#      pattern: "{d} - {m}{n}"
-
-# Set the default logging level to "warn" and attach the "stdout" appender to the root
+    path: "log/stdout.log"
 root:
-  level: debug
+  level: warn
   appenders:
-    - requests
-
-#loggers:
-  # Raise the maximum log level for events sent to the "app::backend::db" logger to "info"
-#  app::backend::db:
-#    level: info
-
-  # Route log events sent to the "app::requests" logger to the "requests" appender,
-  # and *not* the normal appenders installed at the root
-#  app::requests:
-#    level: info
-#    appenders:
-#      - requests
-#    additive: false
+    - stdout
